api-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
